Split apart module functions into their own files
Allows for selective imports on a function-level for users of
non-treeshaking module bundlers
