--- lib/cookie.js
@@ -1,34 +1,0 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-exports.getCookie = getCookie;
-exports.setCookie = setCookie;
-
-var _jsCookie = require('js-cookie');
-
-var _jsCookie2 = _interopRequireDefault(_jsCookie);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-/**
- * Aliases Cookie.get(). See https://github.com/js-cookie/js-cookie#basic-usage for more info.
- *
- * @param  {string} name Name of the cookie
- * @return {string}      Cookie value
- */
-function getCookie(name) {
-  return _jsCookie2.default.get(name);
-}
-
-/**
- * Aliases Cookie.set(). See https://github.com/js-cookie/js-cookie#basic-usage for more info.
- *
- * @param  {string}        name  Name of the cookie
- * @param  {string|object} value Value to set
- * @param  {number}        days  Days to expire
- */
-function setCookie(key, value, days) {
-  _jsCookie2.default.set(key, value, { expires: days });
-}
