--- lib/file.js
@@ -1,32 +1,0 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-exports.extractFileExtensionFromString = extractFileExtensionFromString;
-exports.extractFileExtensionFromUrl = extractFileExtensionFromUrl;
-/**
- * Extracts a file extension from a string, by splitting by dot and taking the last substring.
- *
- * If the file name does not have an extensions, then this method just returns an empty string.
- * @param  {string} fileName File's name
- * @return {string}          File extension
- *
- * Via: http://stackoverflow.com/a/190878/1263876
- */
-function extractFileExtensionFromString(fileName) {
-  var explodedFileName = fileName.split('.');
-  return explodedFileName.length > 1 ? explodedFileName.pop() : '';
-}
-
-/**
- * Extracts a file extension from a url.
- *
- * If the file name does not have an extensions, then this method just returns an empty string.
- * @param  {string} url Url ending in a file name
- * @return {string}     File extension
- */
-function extractFileExtensionFromUrl(url) {
-  var fileName = url.split('/').pop();
-  return extractFileExtensionFromString(fileName);
-}
