--- lib/request.js
@@ -1,26 +1,0 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-    value: true
-});
-exports.default = request;
-
-require('isomorphic-fetch');
-
-/**
- * Global fetch wrapper that adds some basic error handling
- *
- * @param  {string}  url    Url to request
- * @param  {object}  config Fetch config (see https://developer.mozilla.org/en-US/docs/Web/API/GlobalFetch/fetch#Parameters)
- * @return {Promise}        Promise that will resolve with the response if its status was 2xx;
- *                          otherwise rejects with the response
- */
-function request(url, config) {
-    return fetch(url, config).then(function (res) {
-        // If status is not a 2xx, assume it's an error
-        if (!(res.status >= 200 && res.status <= 300)) {
-            throw res;
-        }
-        return res;
-    });
-}
