+++ src/cookie/index.js
@@ -1,0 +1,2 @@
+export getCookie from './get_cookie';
+export setCookie from './set_cookie';
