+++ src/file/extract_file_extension_from_string.js
@@ -7,20 +7,8 @@
  *
  * Via: http://stackoverflow.com/a/190878/1263876
  */
-export function extractFileExtensionFromString(fileName) {
+export default function extractFileExtensionFromString(fileName) {
     const explodedFileName = fileName.split('.');
     return explodedFileName.length > 1 ? explodedFileName.pop()
                                        : '';
-}
-
-/**
- * Extracts a file extension from a url.
- *
- * If the file name does not have an extensions, then this method just returns an empty string.
- * @param  {string} url Url ending in a file name
- * @return {string}     File extension
- */
-export function extractFileExtensionFromUrl(url) {
-    const fileName = url.split('/').pop();
-    return extractFileExtensionFromString(fileName);
 }
