+++ src/file/index.js
@@ -1,0 +1,2 @@
+export extractFileExtensionFromString from './extract_file_extension_from_string';
+export extractFileExtensionFromUrl from './extract_file_extension_from_url';
