+++ src/index.js
@@ -1,10 +1,23 @@
-// Export the general utils as-is without a namespace
-export * from './general';
+// Export the root-level general utilities without a namespace
+export arrayFrom from './array_from';
+export deepMatchObject from './deep_match_object';
+export formatText from './format_text';
+export intersectLists from './intersect_lists';
+export isShallowEqual from './is_shallow_equal';
+export noop from './noop';
+export omitFromObject from './omit_from_object';
+export request from './request';
+export safeInvoke from './safe_invoke';
+export safeMerge from './safe_merge';
+export sanitize from './sanitize';
+export sanitizeList from './sanitize_list';
+export selectFromObject from './select_from_object';
+export sprint from './sprintf';
 
 export * as Cookie from './cookie';
 export * as DOM from './dom';
 export * as FeatureDetection from './feature_detection';
 export * as File from './file';
 export * as Lang from './lang';
-export * as Request from './request';
+export * as Text from './text';
 export * as Url from './url';
