+++ src/is_shallow_equal.js
@@ -1,0 +1,8 @@
+import { default as isShallowEqual } from 'shallow-equals';
+
+
+/**
+ * Checks shallow equality
+ * Re-export of shallow from shallow-equals
+ */
+export default isShallowEqual;
