+++ src/lang/index.js
@@ -1,0 +1,3 @@
+// Ignore error as we may have more exports in the future
+// eslint-disable-next-line import/prefer-default-export
+export getBrowserLang from './get_browser_lang';
