+++ src/noop.js
@@ -1,0 +1,4 @@
+/**
+ * Noop function that can be stuffed into required callback props
+ */
+export default function noop() {}
