+++ src/sprintf.js
@@ -1,0 +1,9 @@
+import { sprintf } from 'sprintf-js';
+
+
+/**
+ * Formats strings similarly to C's sprintf
+ * Re-export of sprintf from sprintf-js
+ * See https://github.com/alexei/sprintf.js for documentation on usage
+ */
+export default sprintf;
